Update form layout and minor design elements
Decription text is bolder and larger to make it stand out more.
Form elements have been spaced out to make it easier to read.
Form background is white to seperate it from the rest of the right
side content. Changed the button background to make it match left-side
image, but working on final colouring. Added log in link for those
who already havea an account.